Backport r993951 by lvsouza from trunk to the 4.3 branch:
kppplogview uses int to count bytes in and out in a session and to sum up bytes
in and out during a month, when any of those variables reaches 2 GB an overflow
happens. This patch changes variable types from 32-bit to 64-bit.
CCBUG: 196839


svn path=/branches/KDE/4.3/kdenetwork/kppp/; revision=1002341

+++ logview/loginfo.cpp
@@ -59,15 +59,15 @@ double LogInfo::totalCosts() const {
   return _total_cost;
 }
 
-int LogInfo::bytesIn() const {
+long int LogInfo::bytesIn() const {
   return _bytes_in;
 }
 
-int LogInfo::bytesOut() const {
+long int LogInfo::bytesOut() const {
   return _bytes_out;
 }
 
-int LogInfo::bytes() const {
+long int LogInfo::bytes() const {
   if(bytesIn() == -1 || bytesOut() == -1)
     return -1;
   else
